refactor(frontend): 拆分 ApiClient 为多个职责单一的客户端类
将 1217 行的单一 api.ts 文件拆分为 8 个职责明确的客户端类：
- ConfigClient: 配置管理 API
- StatusClient: 状态管理 API
- ToolClient: 工具管理 API
- ServiceClient: 服务控制 API
- VersionClient: 版本信息 API
- EndpointClient: 端点管理 API
- MCPServerClient: MCP 服务器管理 API
- MCPToolClient: MCP 工具管理 API

同时提供组合式 ApiClient 保持向后兼容，确保现有代码无需修改。

解决 GitHub Issue #3124：单一 ApiClient 类包含 8 个功能区域和 58 个方法违反单一职责原则
